requirements.txt
faker
numpy
openai
python-dotenv
fastapi
uvicorn
python-multipart
